refactor(rules): merge ResourceIdentifier and TypedResourceIdentifier
