+++ pkg/core/resources/apis/meshmultizoneservice/status_updater.go
@@ -10,8 +10,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 
-	"github.com/kumahq/kuma/pkg/core/kri"
-	meshmzservice_api "github.com/kumahq/kuma/pkg/core/resources/
 	meshmzservice_api "github.com/kumahq/kuma/pkg/core/resources/apis/meshmultizoneservice/api/v1alpha1"
 	meshservice_api "github.com/kumahq/kuma/pkg/core/resources/apis/meshservice/api/v1alpha1"
 	"github.com/kumahq/kuma/pkg/core/resources/manager"
@@ -20,6 +18,7 @@ import (
 	"github.com/kumahq/kuma/pkg/core/user"
 	core_metrics "github.com/kumahq/kuma/pkg/metrics"
 	"github.com/kumahq/kuma/pkg/core/kri"
+	util_time "github.com/kumahq/kuma/pkg/util/time"
 )
 
 type StatusUpdater struct {
