+++ pkg/core/resources/apis/meshservice/api/v1alpha1/helpers.go
@@ -4,12 +4,11 @@ import (
 	"fmt"
 
 	mesh_proto "github.com/kumahq/kuma/api/mesh/v1alpha1"
-	"github.com/kumahq/kuma/pkg/core/kri"
-	"github.com/ku
 	"github.com/kumahq/kuma/pkg/core/resources/apis/core"
 	core_vip "github.com/kumahq/kuma/pkg/core/resources/apis/core/vip"
 	"github.com/kumahq/kuma/pkg/core/kri"
 	"github.com/kumahq/kuma/pkg/util/pointer"
+	xds_types "github.com/kumahq/kuma/pkg/core/xds/types"
 )
 
 func (m *MeshServiceResource) DestinationName(port uint32) string {
