+++ pkg/plugins/policies/core/xds/meshroute/gateway/gateway.go
@@ -9,14 +9,13 @@ import (
 	core_mesh "github.com/kumahq/kuma/pkg/core/resources/apis/mesh"
 	core_xds "github.com/kumahq/kuma/pkg/core/xds"
 	"github.com/kumahq/kuma/pkg/plugins/policies/core/rules"
-	"github.com/kumahq/kuma/pkg/plugins/policies/core/rules/resolve"
-	p
 	plugin_gateway "github.com/kumahq/kuma/pkg/plugins/runtime/gateway"
 	"github.com/kumahq/kuma/pkg/plugins/runtime/gateway/match"
 	util_maps "github.com/kumahq/kuma/pkg/util/maps"
 	xds_context "github.com/kumahq/kuma/pkg/xds/context"
 	"github.com/kumahq/kuma/pkg/plugins/policies/core/rules/resolve"
 	xds_topology "github.com/kumahq/kuma/pkg/xds/topology"
+	envoy_names "github.com/kumahq/kuma/pkg/xds/envoy/names"
 )
 
 type Sublistener struct {
