+++ pkg/plugins/policies/meshtrace/plugin/v1alpha1/plugin.go
@@ -15,8 +15,6 @@ import (
 	xds_types "github.com/kumahq/kuma/pkg/core/xds/types"
 	"github.com/kumahq/kuma/pkg/plugins/policies/core/matchers"
 	core_rules "github.com/kumahq/kuma/pkg/plugins/policies/core/rules"
-	"github.com/kumahq/kuma/pkg/plugins/policies/core/rules/resolve"
-	p
 	policies_xds "github.com/kumahq/kuma/pkg/plugins/policies/core/xds"
 	"github.com/kumahq/kuma/pkg/plugins/policies/core/xds/meshroute"
 	api "github.com/kumahq/kuma/pkg/plugins/policies/meshtrace/api/v1alpha1"
@@ -25,6 +23,7 @@ import (
 	xds_context "github.com/kumahq/kuma/pkg/xds/context"
 	"github.com/kumahq/kuma/pkg/plugins/policies/core/rules/resolve"
 	xds_topology "github.com/kumahq/kuma/pkg/xds/topology"
+	"github.com/kumahq/kuma/pkg/xds/envoy/clusters"
 )
 
 const OriginMeshTrace = "mesh-trace"
